extract runtime atom budget helpers
